+++ apps/docs/components/ProjectConfigVariables/ProjectConfigVariables.tsx
@@ -1,66 +1,31 @@
-import { useMemo, useState } from 'react'
-import { Check, ChevronsUpDown } from 'lucide-react'
-import {
-  Button_Shadcn_ as Button,
-  Popover_Shadcn_ as Popover,
-  PopoverTrigger_Shadcn_ as PopoverTrigger,
-  PopoverContent_Shadcn_ as PopoverContent,
-  Command_Shadcn_ as Command,
-  CommandInput_Shadcn_ as CommandInput,
-  CommandEmpty_Shadcn_ as CommandEmpty,
-  CommandItem_Shadcn_ as CommandItem,
-  CommandGroup_Shadcn_ as CommandGroup,
-  Input_Shadcn_ as Input,
-  cn,
-  IconCopy,
-  IconCheck,
-  ScrollArea,
-} from 'ui'
-import CopyToClipboard from 'react-copy-to-clipboard'
-import { proxy, useSnapshot } from 'valtio'
-import { LOCAL_STORAGE_KEYS, retrieve, store } from '~/lib/storage'
-import Link from 'next/link'
 import { isBrowser, useIsLoggedIn, useIsUserLoading } from 'common'
-import { ProjectsData, useProjectsQuery } from '~/lib/fetch/projects'
-import { OrganizationsData, useOrganizationsQuery } from '~/lib/fetch/organizations'
-import { BranchesData, useBranchesQuery } from '~/lib/fetch/branches'
+import Link from 'next/link'
+import { useMemo } from 'react'
+import CopyToClipboard from 'react-copy-to-clipboard'
+import { Button_Shadcn_ as Button, Input_Shadcn_ as Input, cn, IconCopy, IconCheck } from 'ui'
+import { proxy, useSnapshot } from 'valtio'
+
+import { useCopy } from '~/hooks/useCopy'
+import { useBranchesQuery } from '~/lib/fetch/branches'
+import { useOrganizationsQuery } from '~/lib/fetch/organizations'
 import { useProjectApiQuery } from '~/lib/fetch/projectApi'
+import { useProjectsQuery } from '~/lib/fetch/projects'
+import { LOCAL_STORAGE_KEYS, remove, retrieve, store } from '~/lib/storage'
 import { useOnLogout } from '~/lib/userAuth'
-import { noop } from 'lodash'
 
-type Org = OrganizationsData[number]
-type Project = ProjectsData[number]
-type Branch = BranchesData[number]
-
-const projectsStore = proxy({
-  // Local storage keys are cleared globally at the app level,
-  // so don't need to worry about clearing them here.
-  selectedOrg: null as Org | null,
-  selectedProject: null as Project | null,
-  setSelectedOrgProject: (org: Org | null, project: Project | null) => {
-    projectsStore.selectedOrg = org
-    if (org) store('local', LOCAL_STORAGE_KEYS.SAVED_ORG, org.id)
-    projectsStore.selectedProject = project
-    // @ts-ignore -- problem in OpenAPI spec -- project has ref property
-    if (project) store('local', LOCAL_STORAGE_KEYS.SAVED_PROJECT, project.ref)
-  },
-  selectedBranch: null as Branch | null,
-  setSelectedBranch: (branch: Branch | null) => {
-    projectsStore.selectedBranch = branch
-    if (branch) store('local', LOCAL_STORAGE_KEYS.SAVED_BRANCH, branch.id)
-  },
-  clear: () => {
-    projectsStore.setSelectedOrgProject(null, null)
-    projectsStore.setSelectedBranch(null)
-  },
-})
-
-type Variable = 'url' | 'anonKey'
-
-const prettyFormatVariable: Record<Variable, string> = {
-  url: 'Project URL',
-  anonKey: 'Anon key',
-}
+import { ComboBox, ComboBoxOption } from './ProjectConfigVariables.ComboBox'
+import {
+  Branch,
+  Org,
+  Project,
+  Variable,
+  fromBranchValue,
+  fromOrgProjectValue,
+  prettyFormatVariable,
+  toBranchValue,
+  toDisplayNameOrgProject,
+  toOrgProjectValue,
+} from './ProjectConfigVariables.utils'
 
 type ProjectOrgDataState =
   | 'userLoading'
@@ -70,139 +35,50 @@ type ProjectOrgDataState =
   | 'loggedIn.dataSuccess.hasData'
   | 'loggedIn.dataSuccess.hasNoData'
 
-interface ComboBoxOption {
-  id: string
-  value: string
-  displayName: string
-}
+type BranchesDataState =
+  | 'userLoading'
+  | 'loggedOut'
+  | 'loggedIn.noBranches'
+  | 'loggedIn.branches.dataPending'
+  | 'loggedIn.branches.dataError'
+  | 'loggedIn.branches.dataSuccess.hasData'
+  | 'loggedIn.branches.dataSuccess.noData'
 
-function ComboBox<Opt extends ComboBoxOption>({
-  isLoading,
-  disabled,
-  name,
-  options,
-  selectedOption,
-  onSelectOption = noop,
-  className,
-}: {
-  isLoading: boolean
-  disabled?: boolean
-  name: string
-  options: Opt[]
-  selectedOption?: string
-  onSelectOption?: (newValue: string) => void
-  checkEqualsOptions?: (old: string, candidate: string) => boolean
-  className?: string
-}) {
-  const [open, setOpen] = useState(false)
-
-  return (
-    <Popover open={open} onOpenChange={setOpen}>
-      <div className={cn('flex items-center', className)}>
-        <PopoverTrigger asChild aria-label={`Select ${name}`} disabled={disabled}>
-          <Button
-            variant="outline"
-            role="combobox"
-            aria-expanded={open}
-            className={cn(
-              'overflow-hidden',
-              'h-auto min-h-10',
-              'flex justify-between',
-              'border-none',
-              'py-0 pl-0 pr-1 text-left'
-            )}
-          >
-            {isLoading
-              ? 'Loading...'
-              : (selectedOption && fromOrgProjectValue(selectedOption)[2]) ?? 'Select a project...'}
-            <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-          </Button>
-        </PopoverTrigger>
-      </div>
-      <PopoverContent className="z-[999999] p-0" side="bottom">
-        <Command>
-          <CommandInput placeholder="Search project..." className="border-none ring-0" />
-          <CommandEmpty>No project found.</CommandEmpty>
-          <CommandGroup>
-            <ScrollArea className={options.length > 10 ? 'h-[280px]' : ''}>
-              {options.map((option) => (
-                <CommandItem
-                  key={option.id}
-                  value={option.value}
-                  onSelect={(selectedValue: string) => {
-                    setOpen(false)
-                    onSelectOption(selectedValue)
-                  }}
-                  className="cursor-pointer"
-                >
-                  <Check
-                    className={cn(
-                      'mr-2 h-4 w-4',
-                      selectedOption === option.value ? 'opacity-100' : 'opacity-0'
-                    )}
-                  />
-                  {option.displayName}
-                </CommandItem>
-              ))}
-            </ScrollArea>
-          </CommandGroup>
-        </Command>
-      </PopoverContent>
-    </Popover>
-  )
-}
-
-function useCopy() {
-  const [copied, setCopied] = useState(false)
-
-  function handleCopy() {
-    setCopied(true)
-    setTimeout(() => {
-      setCopied(false)
-    }, 1000)
-  }
-
-  return { copied, handleCopy }
-}
-
-function toDisplayNameOrgProject(org: Org, project: Project) {
-  return `${org.name} / ${project.name}`
-}
-
-function toOrgProjectValue(org: Org, project: Project) {
-  // @ts-ignore -- problem in OpenAPI spec -- project has ref property
-  return JSON.stringify([org.id, project.ref, toDisplayNameOrgProject(org, project)])
-}
-
-function fromOrgProjectValue(
-  maybeOrgProject: string
-): [string, string, string] | [null, null, null] {
-  try {
-    const data = JSON.parse(maybeOrgProject)
-    if (!Array.isArray(data) || data.length !== 3) {
-      throw Error("Shape of parsed JSON doesn't match form of org and project value")
+const projectsStore = proxy({
+  // Local storage keys are cleared globally at the app level,
+  // so don't need to worry about clearing them here.
+  selectedOrg: null as Org | null,
+  selectedProject: null as Project | null,
+  setSelectedOrgProject: (org: Org | null, project: Project | null) => {
+    projectsStore.selectedOrg = org
+    if (org) {
+      store('local', LOCAL_STORAGE_KEYS.SAVED_ORG, org.id)
+    } else {
+      remove('local', LOCAL_STORAGE_KEYS.SAVED_ORG)
     }
-    return [data[0], data[1], data[2]]
-  } catch {
-    return [null, null, null]
-  }
-}
 
-function toBranchValue(branch: Branch) {
-  return JSON.stringify([branch.id, branch.name])
-}
-
-function fromBranchValue(maybeBranch: string): [string, string] | [null, null] {
-  try {
-    const data = JSON.parse(maybeBranch)
-    if (!Array.isArray(data) || data.length !== 2) {
-      throw Error("Shape of parsed JSON doesn't match form of branch value")
+    projectsStore.selectedProject = project
+    if (project) {
+      // @ts-ignore -- problem in OpenAPI spec -- project has ref property
+      store('local', LOCAL_STORAGE_KEYS.SAVED_PROJECT, project.ref)
+    } else {
+      remove('local', LOCAL_STORAGE_KEYS.SAVED_PROJECT)
     }
-    return [data[0], data[1]]
-  } catch {
-    return [null, null]
-  }
-}
+  },
+  selectedBranch: null as Branch | null,
+  setSelectedBranch: (branch: Branch | null) => {
+    projectsStore.selectedBranch = branch
+    if (branch) {
+      store('local', LOCAL_STORAGE_KEYS.SAVED_BRANCH, branch.id)
+    } else {
+      remove('local', LOCAL_STORAGE_KEYS.SAVED_BRANCH)
+    }
+  },
+  clear: () => {
+    projectsStore.setSelectedOrgProject(null, null)
+    projectsStore.setSelectedBranch(null)
+  },
+})
 
 function OrgProjectSelector() {
   const isUserLoading = useIsUserLoading()
@@ -303,21 +179,37 @@ function OrgProjectSelector() {
 }
 
 function BranchSelector() {
+  const userLoading = useIsUserLoading()
   const isLoggedIn = useIsLoggedIn()
+
   const { selectedProject, selectedBranch, setSelectedBranch } = useSnapshot(projectsStore)
 
   // @ts-ignore -- problem in OpenAPI spec -- project has is_branch_enabled property
   const hasBranches = selectedProject?.is_branch_enabled ?? false
   if (!hasBranches) setSelectedBranch(null)
 
-  const { data, isPending, isSuccess } = useBranchesQuery(
+  const { data, isPending, isSuccess, isError } = useBranchesQuery(
     // @ts-ignore -- problem in OpenAPI spec -- project has ref property
     { projectRef: selectedProject?.ref },
     { enabled: isLoggedIn && hasBranches }
   )
 
+  const stateSummary: BranchesDataState = userLoading
+    ? 'userLoading'
+    : !isLoggedIn
+    ? 'loggedOut'
+    : !hasBranches
+    ? 'loggedIn.noBranches'
+    : isPending
+    ? 'loggedIn.branches.dataPending'
+    : isError
+    ? 'loggedIn.branches.dataError'
+    : data.length === 0
+    ? 'loggedIn.branches.dataSuccess.noData'
+    : 'loggedIn.branches.dataSuccess.hasData'
+
   const formattedData: ComboBoxOption[] =
-    !hasBranches || !isSuccess || data.length === 0
+    stateSummary !== 'loggedIn.branches.dataSuccess.hasData'
       ? []
       : data.map((branch) => ({
           id: branch.id,
@@ -325,7 +217,7 @@ function BranchSelector() {
           value: toBranchValue(branch),
         }))
 
-  if (isBrowser && isSuccess && !selectedBranch) {
+  if (isBrowser && stateSummary === 'loggedIn.branches.dataSuccess.hasData' && !selectedBranch) {
     const storedMaybeBranchId = retrieve('local', LOCAL_STORAGE_KEYS.SAVED_BRANCH)
     const storedBranch = data.find((branch) => branch.id === storedMaybeBranchId)
 
@@ -342,8 +234,8 @@ function BranchSelector() {
   return hasBranches ? (
     <ComboBox
       name="branch"
-      isLoading={isPending}
-      disabled={!isPending && data?.length === 0}
+      isLoading={stateSummary === 'userLoading' || stateSummary === 'loggedIn.branches.dataPending'}
+      disabled={stateSummary !== 'loggedIn.branches.dataSuccess.hasData'}
       options={formattedData}
       selectedOption={toBranchValue(selectedBranch)}
       onSelectOption={(option) => {
