+++ apps/docs/components/ProjectConfigVariables/index.tsx
@@ -1,0 +1,1 @@
+export { ProjectConfigVariables } from './ProjectConfigVariables'
